requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
